Added the remaining frog cards

+++ Arkhana/Arkhana/Modifier.h
@@ -10,7 +10,7 @@ public:
 	Modifier(Modifier*);
 	~Modifier() {}
 	int GetValue() { return value; }
-	int GetMultiplier() { return multiplier; }
+	float GetMultiplier() { return multiplier; }
 	STAT_TYPE GetStat() { return sType; }
 	MODIFIER_TYPE GetModType() { return mType; }
 	string GetText() { return text; }
@@ -68,7 +68,7 @@ protected:
 	bool bActive = true;
 	int value = 1;
 	int duration = 999;
-	int multiplier = 0;
+	float multiplier = 0;
 	int EOTChange = 0;
 	STAT_TYPE sType = STAT_TYPE::DMG_PHYSICAL;
 	MODIFIER_TYPE mType = MODIFIER_TYPE::ZONE_MOD;
